+++ internal/raft/logentry_etcd_test.go
@@ -716,13 +716,13 @@ func TestCompactionSideEffects(t *testing.T) {
 	lastTerm := lastIndex
 	storage := NewTestLogDB()
 	for i = 1; i <= unstableIndex; i++ {
-		if err := storage.Append([]pb.Entry{{Term: uint64(i), Index: uint64(i)}}); err != nil {
+		if err := storage.Append([]pb.Entry{{Term: i, Index: i}}); err != nil {
 			t.Fatalf("%v", err)
 		}
 	}
 	raftLog := newEntryLog(storage, server.NewInMemRateLimiter(0))
 	for i = unstableIndex; i < lastIndex; i++ {
-		raftLog.append([]pb.Entry{{Term: uint64(i + 1), Index: uint64(i + 1)}})
+		raftLog.append([]pb.Entry{{Term: i + 1, Index: i + 1}})
 	}
 
 	ok := raftLog.tryCommit(lastIndex, lastTerm)
